src/django/requirements.txt
Django==2.0.9
django-jsonview==1.2.0
django-watchman==0.15.0
django-extensions==2.1.4
djangorestframework==3.9.0
djangorestframework-gis==0.14
django-rest-auth==0.9.3
django-allauth==0.37.0
flake8==3.6.0
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
